modification des proprietes de la classe polygon, du type double en float

+++ src/main/java/com/yvanscoop/apicoordonneesville/entities/BoundingBox.java
@@ -9,8 +9,8 @@ import java.io.Serializable;
 @Embeddable
 @Getter @Setter
 public class BoundingBox implements Serializable {
-    private Double southLatitude;
-    private Double northLatitude;
-    private Double westLongitude;
-    private Double eastLongitude;
+    private float southLatitude;
+    private float northLatitude;
+    private float westLongitude;
+    private float eastLongitude;
 }
